Create file to handle operations with config files
Refactor device source file to separate, in a new file, functions that
interact with configurations files.
