+++ Makefile.am
@@ -22,7 +22,8 @@ src_thingd_SOURCES = src/main.c \
 			src/iface-modbus.c src/iface-modbus.h \
 			src/settings.c src/settings.h \
 			src/knot-config.c src/knot-config.h \
-			src/poll.c src/poll.h
+			src/poll.c src/poll.h \
+			src/properties.c src/properties.h
 
 src_thingd_LDADD = $(modules_ldadd) -lm
 src_thingd_LDFLAGS = $(AM_LDFLAGS)
