+++ src/sm.c
@@ -21,6 +21,7 @@
 
 #include <stdlib.h>
 #include <errno.h>
+#include <knot/knot_protocol.h>
 #include <ell/ell.h>
 
 #include "sm-pvt.h"
